+++ spec/slather/project_spec.rb
@@ -53,14 +53,14 @@ describe Slather::Project do
     it "should return coverage file objects of type coverage_file_class for unignored project files" do
       fixtures_project.ignore_list = ["*fixturesTests*"]
       allow(fixtures_project).to receive(:dedupe) { |coverage_files| coverage_files }
-      coverage_files = fixtures_project.send(:coverage_files)
+      coverage_files = fixtures_project.coverage_files
       coverage_files.each { |cf| expect(cf.kind_of?(SpecCoverageFile)).to be_truthy }
       expect(coverage_files.map { |cf| cf.source_file_pathname.basename.to_s }).to eq(["fixtures.m", "peekaview.m"])
     end
 
     it "should raise an exception if no unignored project coverage file files were found" do
       fixtures_project.ignore_list = ["*fixturesTests*", "*fixtures*"]
-      expect {fixtures_project.send(:coverage_files)}.to raise_error(StandardError)
+      expect {fixtures_project.coverage_files}.to raise_error(StandardError)
     end
   end